Move Datatables processing into a helper
